Interface change in ExtendTrajectoryToNeighboringUnitCells
Instead of having to pass a list of vectors for each frame in the trajectory
which is used to extend the simulation, we can now simply pass a list
containing the directions in which the trajectory should be extended.
A new function ExtendTrajectory has been created to maintain the possibility of directly
passing the list of vectors used for extension.